Refactor: relocate Verify.Problem to its own class
