+++ src/main/java/edu/tamu/di/SAFCreator/ImportDataProcessorImpl.java
@@ -32,10 +32,9 @@ import edu.tamu.di.SAFCreator.model.FileLabel;
 import edu.tamu.di.SAFCreator.model.FlagPanel;
 import edu.tamu.di.SAFCreator.model.HandleLabel;
 import edu.tamu.di.SAFCreator.model.Item;
+import edu.tamu.di.SAFCreator.model.Problem;
 import edu.tamu.di.SAFCreator.model.SchematicFieldSet;
 import edu.tamu.di.SAFCreator.model.StubLabel;
-import edu.tamu.di.SAFCreator.model.verify.Verifier;
-import edu.tamu.di.SAFCreator.model.verify.Verifier.Problem;
 
 public class ImportDataProcessorImpl implements ImportDataProcessor {
     private static String PdfPrefix = "document-";
@@ -366,7 +365,7 @@ public class ImportDataProcessorImpl implements ImportDataProcessor {
                 method = this.getClass().getMethod("isCancelled");
                 problems = item.writeItemSAF(this, method);
 
-                for (Verifier.Problem problem : problems) {
+                for (Problem problem : problems) {
                     console.append("\t" + problem.toString() + "\n");
                     if (problem.isError()) {
                         hasError = true;
