+++ src/main/java/edu/tamu/di/SAFCreator/model/verify/Verifier.java
@@ -5,72 +5,10 @@ import java.util.List;
 import javax.swing.JTextArea;
 
 import edu.tamu.di.SAFCreator.model.Batch;
-import edu.tamu.di.SAFCreator.model.Flag;
 import edu.tamu.di.SAFCreator.model.FlagPanel;
+import edu.tamu.di.SAFCreator.model.Problem;
 
 public interface Verifier extends VerifierProperty {
-
-    public class Problem {
-        private Integer rownumber = null;
-        private String columnLabel = null;
-        private boolean error;
-        private String note;
-        private Flag flag = null;
-
-        public Problem(boolean error, String note) {
-            rownumber = null;
-            columnLabel = null;
-            this.error = error;
-            this.note = note;
-            flag = null;
-        }
-
-        public Problem(int rownumber, String columnLabel, boolean error, String note) {
-            this.rownumber = rownumber;
-            this.columnLabel = columnLabel;
-            this.error = error;
-            this.note = note;
-            flag = null;
-        }
-
-        public Problem(int rownumber, String columnLabel, boolean error, String note, Flag flag) {
-            this.rownumber = rownumber;
-            this.columnLabel = columnLabel;
-            this.error = error;
-            this.note = note;
-            this.flag = flag;
-        }
-
-        public Flag getFlag() {
-            return flag;
-        }
-
-        public boolean isError() {
-            return error;
-        }
-
-        public boolean isFlagged() {
-            return flag != null;
-        }
-
-        public void setFlag(Flag flag) {
-            this.flag = flag;
-        }
-
-        @Override
-        public String toString() {
-            String flagged = "";
-            if (flag != null) {
-                flagged = "Flagged ";
-            }
-            if (rownumber == null) {
-                return flagged + (error ? "ERROR" : "WARNING") + ": " + note;
-            }
-
-            return flagged + (error ? "ERROR at " : "WARNING at ") + "column " + columnLabel + " row " + rownumber
-                    + ":\n\t" + note;
-        }
-    }
 
     public List<Problem> verify(Batch batch);
