+++ src/main/java/edu/tamu/di/SAFCreator/model/verify/impl/ValidSchemaNameVerifierImpl.java
@@ -11,6 +11,7 @@ import edu.tamu.di.SAFCreator.model.ColumnLabel;
 import edu.tamu.di.SAFCreator.model.FieldLabel;
 import edu.tamu.di.SAFCreator.model.Flag;
 import edu.tamu.di.SAFCreator.model.FlagPanel;
+import edu.tamu.di.SAFCreator.model.Problem;
 import edu.tamu.di.SAFCreator.model.verify.VerifierBackground;
 import edu.tamu.di.SAFCreator.model.verify.VerifierProperty;
